Add files via upload
This python script is designed to be used as a hotkey.  It will flip the view of the camera left to right to give the user fresh eyes on their work.  You need to have the desired camera selected the first time you press the hotkey.  It will then create a camera and group called 'flipCamGrp'.  After that pressing the hotkey will flip the camera back and forth.  Manually delete the group from the outliner to start over.